refactor(issue): move sidebar hooks to wrapper and simplify detail page
moved hooks for sidebar data and edits out of the issue detail page into a dedicated wrapper to reduce bloat.
ensured consistent use of issueId as key. moved toggle and edit hooks for issue status into the header component.
